refactor(all/api): api code splitting
