+++ src/renderer/api/nt/nt.d.ts
